Tales.Vesperia.Website: Write out NPC stuff I guess.
This was intended as a cross-comparison between X360 and PS3 NPC
dialogue text, but it turns out -- and I should have realized this
sooner, considering I had to do stuff to transfer over English NPC
dialogue from X360 to PS3 -- that the two versions store NPC dialogue
references in a completely different and incompatible manner.
Unfortunate.