Tales.Vesperia.Website: Export pre-made website data to database.
Frankly, this is very ugly, but I tried it the "proper" way and quickly
realized that not only my old server is too crappy to generate full
pages of several hundred items or artes with seveal references to
enemies or locations or whatever in acceptable timeframes, but I'd also
have to effectively rewrite all exporting code in PHP. So instead, just
store the pre-made HTML for each item/arte/enemy/... in the DB, and show
that. Not pretty, but fast and accurate.