Experimental PartialStream.
This is intended to get Streams from file containers that can be used just like actual FileStreams to files on disk without having to actually extract the files first.

For somewhat sensible usage and the ability to have several Streams within the same container file, we need a way to clone an existing Stream with little overhead, which is the task of the DuplicatableStream abstract class.

This is a bit experimental, but should result in a neat way to traverse file archives without having to extract them first.